% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{as.data.frame.spgraph}
\alias{as.data.frame.spgraph}
\title{This function creates a data frame out of the shortest_path_graphs in order to show the optimization steps}
\usage{
\method{as.data.frame}{spgraph}(spgraph)
}
\arguments{
\item{\code{spgraph}}{object}
}
\value{
If Dijkstra :The data frame consisting of mininum distance from source node, shortest path predecessor and front.
        If FloydWarshall :The data frame consisting of mininum distance between all nodes and their shortest path predecessor.
}
\description{
This function creates a data frame out of the shortest_path_graphs in order to show the optimization steps
}

