% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spgraph.R
\name{as.spgraph}
\alias{as.spgraph}
\alias{as.spgraph.igraph}
\title{Coerce to a shortest path graph}
\usage{
as.spgraph(x)

\method{as.spgraph}{igraph}(x)
}
\arguments{
\item{x}{Object to coerce}
}
\description{
This method can be used to convert objects into spgraphs.
}
\section{Methods (by class)}{
\itemize{
\item \code{igraph}: Convert an igraph to an spgraph
}}

