% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bellmanFord.R
\name{bellmanFord}
\alias{bellmanFord}
\title{Calculates the shortest path from a source vertex to a target vertex in a graph}
\usage{
bellmanFord(graph, from, to)
}
\arguments{
\item{graph}{The \code{igraph} object.}

\item{from}{Source node}

\item{to}{Target node}
}
\description{
Calculates the shortest path from a source vertex to a target vertex in a graph
}

