% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bellmanFordRcpp}
\alias{bellmanFordRcpp}
\title{Bellman-Ford's Algorithm to find shortest path from a node to others}
\usage{
bellmanFordRcpp(matriceAdjacence, source)
}
\arguments{
\item{matriceAdjacence}{Data.frame object with three variables (from, to and weight)
that contains the edges of the graph_DataFrame (from, to) with the weight of the edge (w)}

\item{source}{initial node (could be int or str).}
}
\value{
list of two elements.
Distance: the shortest distance from the start node to other nodes (vector)
parents : the predecessor of each node (vector)
}
\description{
Bellman-Ford's Algorithm to find shortest path from a node to others
}
\examples{
matriceAdjacence = t(matrix(data = c(0  , 0  ,  0  ,  0  ,  0  ,    5  , 0,
                                   3  , 0  ,  0  ,  10 ,  3  ,   11  , 0,
                                   0  , 1  ,  0  ,  7  ,  0  ,   0   , 0,
                                   0  , 0  ,  0  ,  0  ,   0 ,    0  , 0,
                                   0  , 0  ,  0  ,  4  ,  0  ,    0  , 0,
                                   1  , 0  ,  0  ,  0  ,  0  ,    0  , 0,
                                   5  , 0  ,  0  ,  0  ,  4  ,    0  , 0), nrow  = 7))
bellmanFordRcpp(matriceAdjacence,2)

}
