% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{euclidean.vertex.distance}
\alias{euclidean.vertex.distance}
\title{Euclidean distance computation for vertices with \code{x} and \code{y} attributes.}
\usage{
euclidean.vertex.distance(graph, v1, v2)
}
\arguments{
\item{graph}{The igraph object}

\item{v1}{The first vertice vector}

\item{v2}{The second vertice vector.}
}
\description{
Euclidean distance computation for vertices with \code{x} and \code{y} attributes.
}

