% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.vertex}
\alias{get.vertex}
\title{get.vertex}
\usage{
get.vertex(graph, identifier)
}
\arguments{
\item{graph}{The igraph object}

\item{identifier}{the identifier}
}
\description{
Convenience function that returns the vertex object associated with the
given identifier, which is either the vertex object itself or a vertex id.
}

