% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getShortestPaths.R
\name{getShortestPaths}
\alias{getShortestPaths}
\alias{getShortestPaths.spgraph}
\alias{getShortestPaths.spresults}
\title{Get the shortest path between two nodes.}
\usage{
getShortestPaths(x, from = x$from, to = x$to)

\method{getShortestPaths}{spresults}(x, from = x$from, to = x$to)

\method{getShortestPaths}{spgraph}(x, from = x$from, to = x$to)
}
\arguments{
\item{x}{A spgraph or a spresults object.}

\item{from}{Source node}

\item{to}{Target node}

\item{...}{Additional arguments passed to \code{getShortestPaths.spgraph}}
}
\description{
Get the shortest path between two nodes.
}
\section{Methods (by class)}{
\itemize{
\item \code{spresults}: Get the shortest path given final spgraph.

\item \code{spgraph}: Get the shortest path given the
knowledge of the provided spgraph instance.
}}

