% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{has.vertex.coordinates}
\alias{has.vertex.coordinates}
\title{Returns \code{TRUE} if the graph's vertices have non-null coordinates.}
\usage{
has.vertex.coordinates(graph)
}
\arguments{
\item{graph}{The graph to check}
}
\description{
Returns \code{TRUE} if the graph's vertices have non-null coordinates.
}

