% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.edge.intersection}
\alias{is.edge.intersection}
\title{is.edge.intersection}
\usage{
is.edge.intersection(g, edges_or_vertices)
}
\arguments{
\item{g}{The graph object}

\item{edges_or_vertices}{Either a vector containing two edges or a matrix
containing four vertices (as edge endpoints).}
}
\value{
\code{TRUE} if both edges intersect, \code{FALSE} otherwise.
}
\description{
Determine if two edges intersect.
}
