% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spresults.R
\name{is.spresults}
\alias{is.spresults}
\title{Check whether an object is a shortest path result object}
\usage{
is.spresults(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if its argument is a shortest path result object, \code{FALSE} otherwise.
}
\description{
Check whether an object is a shortest path result object
}

