% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{nice_vertex_label_colors}
\alias{nice_vertex_label_colors}
\title{Produce a vector of suitable vector colors depending on the vertex' state.}
\usage{
nice_vertex_label_colors(graph, previous = NULL, colors = c("black",
  "darkgray", wes_palette("Royal1")[2]))
}
\arguments{
\item{graph}{The spgraph object.}

\item{previous}{The spgraph object of the last iteration. If provided, changed min_dists
will be highlighted.}

\item{colors}{A \code{c(normal, infinite distance, highlight)} color tuple}
}
\description{
Produce a vector of suitable vector colors depending on the vertex' state.
}

