% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.spgraph}
\alias{plot.spgraph}
\title{Plotting of shortest path graphs}
\usage{
\method{plot}{spgraph}(x, vertex.color = wes_palette("Royal1")[c(3, 4, 1)],
  vertex.color.by = c("set", "manual"),
  vertex.frame.color = wes_palette("Rushmore")[c(4, 3, 5)],
  vertex.frame.color.by = c("type", "manual"), edge.label = "weight",
  vertex.label = "auto", ...)
}
\arguments{
\item{x}{The graph to plot.}

\item{vertex.color}{The vertex coloring. Passing 'set' will color,
the vertices according to the set they are in, other values are passed to
igraph as-is.}

\item{edge.label}{The edge label. Passing 'weight' will print the edge's weight.}

\item{vertex.label}{The vertex label. Passing 'auto' will print the vertex's name
and its currently known minimal distance.}

\item{...}{All other parameters are passed to \code{\link{plot.igraph}} as is.}
}
\description{
\code{plot.spgraph} is a wrapper around \code{\link{plot.igraph}} providing
convenience features for shortest path graph plotting.
}
\examples{
# TODO
}

