% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomGraph.R
\name{randomGraph}
\alias{randomGraph}
\title{Random Graph Generation}
\usage{
randomGraph(no.of.nodes = 12, k = 2.5, euclidean = FALSE)
}
\arguments{
\item{no.of.nodes}{Integer scalar, the number of vertices in the generated
graph.}

\item{k}{The average degree of each vertex in the graph}

\item{euclidean}{Set to \code{TRUE} if (rounded) euclidean distances should be used.}
}
\description{
This method generates a random graph suitable for exercises.
}

