% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomGraph.R
\name{sample_average_k_connected_graph}
\alias{sample_average_k_connected_graph}
\title{Create a random regular graph}
\usage{
sample_average_k_connected_graph(no.of.nodes, k)
}
\arguments{
\item{no.of.nodes}{Integer scalar, the number of vertices in the generated
graph.}

\item{k}{The average degree of each vertex in the graph}
}
\description{
Generate a "nice" random graph with a fixed average vertex degree
and at most one cluster.
}

