\name{shortestpath-package}
\alias{shortestpath-package}
\alias{shortestpath}
\docType{package}
\title{
\packageTitle{shortestpath}
}
\description{
\packageDescription{shortestpath}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{shortestpath}
\packageIndices{shortestpath}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{shortestpath}

Maintainer: \packageMaintainer{shortestpath}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
