% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortestPathAlgo-package.R
\docType{package}
\name{shortestPathAlgo-package}
\alias{shortestPathAlgo}
\alias{shortestPathAlgo-package}
\title{shortestPathAlgo: shortestPathAlgo package}
\description{
shortestPathAlgo provide some algorithms to find the sortest path between two nodes in graph
}
\keyword{internal}
