% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spresults.R
\name{spresults}
\alias{spresults}
\title{Convert a list of graphs into an spresults object.}
\usage{
spresults(lst = list())
}
\arguments{
\item{lst}{A list of spgraphs}
}
\description{
Convert a list of graphs into an spresults object.
}

