% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spgraph.R
\name{summary.spgraph}
\alias{summary.spgraph}
\title{summary.spgraph}
\usage{
\method{summary}{spgraph}(object, ...)
}
\arguments{
\item{object}{The graph object which should be summarized}

\item{...}{Further arguments passed to \code{\link[igraph]{print.igraph}}}
}
\description{
summary.spgraph
}

