% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toLatexGraph.R
\name{toLatexGraph}
\alias{toLatexGraph}
\title{Export spgraphs as human-readable TikZ graphs.}
\usage{
toLatexGraph(x)
}
\arguments{
\item{x}{The graph to export.}
}
\description{
This method focuses on a minimal TikZ output that provides a base
for manual modification. For extensive plot customization, see
\code{\link{plot.spgraph}}.
}

