% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toLatexTable.R
\name{toLatexTable}
\alias{toLatexTable}
\title{toLatexTable}
\usage{
toLatexTable(x, ...)
}
\arguments{
\item{x}{The \code{spgraph} object}

\item{...}{Additional arguments passed to \code{getShortestPaths.spgraph}}
}
\description{
Use the toLatexTable function to create a LaTeX code out of a single \code{spgraph}
or out of a \code{spresults} list
}
\examples{
g <- randomGraph(6)
d <- dijkstra(g, "A", "F")

toLatexTable(d)

}

