% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spgraph.R
\name{as.spgraph}
\alias{as.spgraph}
\alias{as.spgraph.igraph}
\title{Coerce to a shortest path graph}
\usage{
as.spgraph(x)

\method{as.spgraph}{igraph}(x)
}
\arguments{
\item{x}{Object to coerce}
}
\description{
You usually do not need to call this function directly.
Instead, supply an igraph to one of the algorithms, which then outputs suitable spgraph objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{igraph}: Convert an igraph to an spgraph
}}

