% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Calculates the shortest path from a starting vertex to an ending vertex}
\usage{
dijkstra(graph, from, to)
}
\arguments{
\item{graph}{The \code{igraph} object.}

\item{from}{The starting vertex from which the shortest path need to be calculated.}

\item{to}{The ending vertex which wants to be reached from the starting vertex by the shortest path.}
}
\value{
A list of \code{spgraph} objects. Each \code{spgraph} object contains information about 
a certain step in the optimization process representing by its attributes
}
\description{
Calculates the shortest path from a starting vertex to an ending vertex
}

