% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/floydWarshall.R
\name{floydWarshall}
\alias{floydWarshall}
\title{Calculates the shortest path of all pairs of vertices in a graph}
\usage{
floydWarshall(graph, weight.attr = "weight")
}
\arguments{
\item{graph}{The \code{igraph} object.}

\item{weight.attr}{Either \code{NULL} or a character string giving an edge attribute name for the edge cost.}
}
\value{
A list of \code{spgraph} objects. Each \code{spgraph} object contains information about
a certain step in the optimization process representing by its attributes
}
\description{
Calculates the shortest path of all pairs of vertices in a graph
}

