% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShortestPaths.R
\name{getShortestPaths}
\alias{getShortestPaths}
\alias{getShortestPaths.spgraph}
\alias{getShortestPaths.spresults}
\title{Get the shortest path between two vertices.}
\usage{
getShortestPaths(x, from = x$from, to = x$to)

\method{getShortestPaths}{spresults}(x, from = x$from, to = x$to)

\method{getShortestPaths}{spgraph}(x, from = x$from, to = x$to)
}
\arguments{
\item{x}{A spgraph or a spresults object.}

\item{from}{the source vertex}

\item{to}{the target vertex}
}
\description{
Get the shortest path between two vertices.
}
\section{Methods (by class)}{
\itemize{
\item \code{spresults}: Get the shortest path given final spgraph.

\item \code{spgraph}: Get the shortest path given the
knowledge of the provided spgraph instance.
}}

