% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spgraph.R
\name{is.spgraph}
\alias{is.spgraph}
\title{Check whether an object is a shortest path graph}
\usage{
is.spgraph(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if its argument is a shortest path graph, \code{FALSE} otherwise.
}
\description{
Check whether an object is a shortest path graph
}

