% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.spgraph}
\alias{is.spgraph}
\title{Plotting of shortest path graphs}
\usage{
is.spgraph(graph)
}
\arguments{
\item{graph}{The graph object.}
}
\value{
\code{TRUE} if its argument is a shortest path graph, \code{FALSE} otherwise.
}
\description{
Plotting of shortest path graphs
}

