% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeShortestPathGraph.R
\name{makeShortestPathGraph}
\alias{makeShortestPathGraph}
\title{Convert an igraph into a shortest path graph}
\usage{
makeShortestPathGraph(graph, singleSource = TRUE)
}
\value{
The \code{spgraph} object.
}
\description{
You usually do not need to call this function directly.
Instead, supply an igraph to one of the algorithms, which then outputs suitable spgraph objects.
}

