% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{nice_vertex_labels}
\alias{nice_vertex_labels}
\title{Produce "name (current min dist)" labels for all vertices.}
\usage{
nice_vertex_labels(graph)
}
\arguments{
\item{graph}{The spgraph object.}
}
\description{
Produce "name (current min dist)" labels for all vertices.
}

