% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.spgraph}
\alias{plot.spgraph}
\title{Plotting of shortest path graphs}
\usage{
\method{plot}{spgraph}(x, vertex.color.by = c("set", "type", "manual"),
  vertex.color = wes_palette("Royal1")[c(3, 4, 1)],
  vertex.frame.color.by = c("type", "set", "manual"),
  vertex.frame.color = wes_palette("Rushmore")[3:5],
  vertex.size.by = c("type", "set", "manual"), vertex.size = c(15, 25, 25),
  edge.color.by = c("shortestpath", "manual"), edge.color = c("darkgrey",
  wes_palette("Darjeeling")[4]), vertex.label = nice_vertex_labels(x),
  edge.label = E(graph)$weight, edge.label.family = "sans",
  edge.label.color = "#444444", edge.label.cex = 0.75,
  vertex.label.color = nice_vertex_label_colors(x),
  vertex.label.dist = 1.25, vertex.label.cex = 0.8,
  vertex.label.family = "sans", vertex.label.degree = -pi/7,
  default.margins = TRUE, ...)
}
\arguments{
\item{x}{The graph to plot.}

\item{vertex.color.by}{Characteristic which should be used to color vertices}

\item{vertex.color}{Color palette for the vertices}

\item{vertex.frame.color.by}{Characteristic which should be used to color vertex frames}

\item{vertex.frame.color}{Color palette for the vertices' frame.}

\item{vertex.size.by}{Characteristic which should be used to determine vertex sizes}

\item{vertex.size}{The size options}

\item{edge.color.by}{Characteristic which should be used to color edges}

\item{edge.color}{Color palette for the edges}

\item{vertex.label}{The vertex label.}

\item{edge.label}{The edge label.}

\item{edge.label.family}{See \code{\link{igraph.plotting}}}

\item{edge.label.color}{See \code{\link{igraph.plotting}}}

\item{edge.label.cex}{See \code{\link{igraph.plotting}}}

\item{vertex.label.color}{See \code{\link{igraph.plotting}},}

\item{vertex.label.dist}{See \code{\link{igraph.plotting}}}

\item{vertex.label.cex}{See \code{\link{igraph.plotting}}}

\item{vertex.label.family}{See \code{\link{igraph.plotting}}}

\item{vertex.label.degree}{See \code{\link{igraph.plotting}}}

\item{default.margins}{If \code{TRUE}, plot margins will be zeroed.}

\item{...}{All other parameters are passed to \code{\link{plot.igraph}} as is.}
}
\description{
\code{plot.spgraph} is a wrapper around \code{\link{plot.igraph}} providing
convenience features for shortest path graph plotting.
}
\details{
Vertex/edge visuals depending on the underlying graph characteristics:

\describe{
 \item{\code{vertex.color.by}}{ denotes which attribute should be used to set the vertex color}
 \item{\code{vertex.color}}{ specifies the color palette, or, if \code{.by} is \code{"manual"}, a color for each vertex.}
}
Available vertex characteristics:
\itemize{
 \item{\code{set}}{ The set to which the vertex belongs: processed/front/unknown}
 \item{\code{type}}{ start node/normal/destination node}
 \item{\code{manual}}{ A color for each vertex is provided manually}
}

Available edge characteristics:
\itemize{
 \item{\code{shortestpath}}{ edge is not on the shortest path / edge is on the shortest path}
 \item{\code{manual}}{ A color for each edge is provided manualy}
}
}
\examples{
g <- randomGraph()
d <- dijkstra(g, "A", "B")
plot(d)

}

