% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{toLatexTable}
\alias{toLatexTable}
\title{This function creates the LaTeX code out of a given object}
\usage{
toLatexTable(x, ...)
}
\arguments{
\item{x}{The \code{spgraph} object}

\item{title}{Title of the latex table}

\item{includeCommand}{blabala}
}
\description{
This function creates the LaTeX code out of a given object
}

