% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-table.R
\name{toLatexTable}
\alias{toLatexTable}
\title{This function creates the LaTex code out of the data frame}
\usage{
toLatexTable(dataFrame, title = "")
}
\arguments{
\item{The}{data frame}
}
\value{
Latex table
}
\description{
This function creates the LaTex code out of the data frame
}

