% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{toLatexTable.spgraph}
\alias{toLatexTable.spgraph}
\title{This function creates the LaTex code out of the data frame}
\usage{
\method{toLatexTable}{spgraph}(spgraph, title = "", includeCommand = TRUE)
}
\arguments{
\item{The}{data frame}
}
\value{
Latex table
}
\description{
This function creates the LaTex code out of the data frame
}

