% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{toLatexTable.spgraph}
\alias{toLatexTable.spgraph}
\title{This function creates the LaTex code out of an spgraph}
\usage{
\method{toLatexTable}{spgraph}(spgraph, title = "", includeCommand = TRUE)
}
\arguments{
\item{spgraph}{The \code{spgraph} object}

\item{title}{Title of the latex table}

\item{includeCommand}{blabala}
}
\value{
Latex table
}
\description{
This function creates the LaTex code out of an spgraph
}

